Split out getTheme
