+++ client/state/themes/selectors/index.js
@@ -31,28 +31,11 @@ import {
 } from 'state/themes/utils';
 import { DEFAULT_THEME_QUERY } from 'state/themes/constants';
 import { FEATURE_UNLIMITED_PREMIUM_THEMES } from 'lib/plans/constants';
+import { getTheme } from 'state/themes/selectors/get-theme';
 
 import 'state/themes/init';
 
-/**
- * Returns a theme object by site ID, theme ID pair.
- *
- * @param  {object}  state   Global state tree
- * @param  {number}  siteId  Site ID
- * @param  {string}  themeId Theme ID
- * @returns {?object}         Theme object
- */
-export const getTheme = createSelector(
-	( state, siteId, themeId ) => {
-		const manager = state.themes.queries[ siteId ];
-		if ( ! manager ) {
-			return null;
-		}
-
-		return manager.getItem( themeId );
-	},
-	state => state.themes.queries
-);
+export { getTheme } from 'state/themes/selectors/get-theme';
 
 /**
  * Returns a theme object from what is considered the 'canonical' source, i.e.
